delay_us() and delay_ms() is ready for STM32F1

+++ ldmicro/LIBRARIES_FOR/Arm/STM32F1/Lib_usr.h
@@ -17,9 +17,9 @@ int16_t  opposite(int16_t var);
 uint8_t bcd2bin(uint8_t var);
 uint8_t bin2bcd(uint8_t var);
 
-extern volatile uint32_t usSysTick;
+extern volatile uint16_t usSysTick;
 
-void delay_us(uint32_t us);
-void delay_ms(uint32_t ms);
+void delay_us(uint16_t us);
+void delay_ms(uint16_t ms);
 
 #endif
